[#2] refactor: handleDeleteTodo로 함수 변경 / type.ts 작성 / getCurrentDate 함수 추가
