+++ src/utils/getCurrentDate.ts
@@ -1,0 +1,5 @@
+import { dateCountry, dateOptions } from 'config';
+
+export const getCurrentDate = () => {
+  return new Date().toLocaleString(dateCountry, dateOptions);
+};
